// src/models/IoK8sApiNodeV1RuntimeClassList/types.ts
import { IoK8sApiNodeV1RuntimeClassModelOptions as IoK8sApiNodeV1RuntimeClass } from '../IoK8sApiNodeV1RuntimeClass';
import { IoK8sApimachineryPkgApisMetaV1ListMetaModelOptions as IoK8sApimachineryPkgApisMetaV1ListMeta } from '../IoK8sApimachineryPkgApisMetaV1ListMeta';

export interface IoK8sApiNodeV1RuntimeClassListModelOptions {
  apiVersion?: string;
  items: Array<IoK8sApiNodeV1RuntimeClass>;
  kind?: string;
  metadata?: IoK8sApimachineryPkgApisMetaV1ListMeta;
}